requirements_arabic_ocr.txt
pandas>=2.2.0
pyarrow>=15.0.0
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
pillow>=10.0.0
